requirements.txt
# --- Core Backend ---
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1

# --- Database / ORM ---
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# --- Auth / Security ---
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# --- API Utilities ---
slowapi==0.1.9
python-multipart==0.0.9

# --- Testing ---
pytest==8.3.3

# --- Document Processing ---
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13

# --- Lightweight AI/NLP ---
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
sentence-transformers==3.0.1
